requirements.txt
awpy==1.0.1
python-telegram-bot==13.11
pandas~=1.4.1
requests==2.27.1